+++ packages/mgt-chat/src/statefulClient/useGraphChatClient.ts
@@ -6,19 +6,8 @@
  */
 
 import { useEffect, useState } from 'react';
-import { v4 as uuid } from 'uuid';
 import { StatefulGraphChatClient } from './StatefulGraphChatClient';
 import { log } from '@microsoft/mgt-element';
-
-/**
- * Provides a stable sessionId for the lifetime of the browser tab.
- * @returns a string that is either read from session storage or generated and placed in session storage
- */
-const useSessionId = (): string => {
-  const [sessionId] = useState<string>(() => uuid());
-
-  return sessionId;
-};
 
 /**
  * Custom hook to abstract the creation of a stateful graph chat client.
@@ -27,27 +16,8 @@ const useSessionId = (): string => {
  * @param {React.EffectCallback} post an optional useEffect function to be run after other useEffects
  * @returns {StatefulGraphChatClient} a stateful graph chat client that is subscribed to the given chatId
  */
-export const useGraphChatClient = (
-  chatId: string,
-  pre: React.EffectCallback | undefined = undefined,
-  post: React.EffectCallback | undefined = undefined
-): StatefulGraphChatClient => {
-  const sessionId = useSessionId();
+export const useGraphChatClient = (): StatefulGraphChatClient => {
   const [chatClient] = useState<StatefulGraphChatClient>(() => new StatefulGraphChatClient());
-
-  // pre
-  // NOTE: we need the registration of state handlers before changing state
-  useEffect(() => {
-    if (pre) pre();
-  }, [chatClient, pre]);
-
-  // when chatId or sessionId changes this effect subscribes or unsubscribes
-  // the component to/from web socket based notifications for the given chatId
-  useEffect(() => {
-    // we must have both a chatId & sessionId to subscribe.
-    if (chatId && sessionId) chatClient.subscribeToChat(chatId, sessionId);
-    else chatClient.setStatus('no chat id');
-  }, [chatId, sessionId, chatClient]);
 
   // Returns a cleanup function to call tearDown on the chatClient
   // This allows us to clean up when the consuming component is being unmounted from the DOM
@@ -57,11 +27,6 @@ export const useGraphChatClient = (
       chatClient.tearDown();
     };
   }, [chatClient]);
-
-  // post
-  useEffect(() => {
-    if (post) post();
-  }, [chatClient, post]);
 
   return chatClient;
 };
